src/app/requirements.txt
Pillow==8.3.2
fastapi==0.70.0